+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `laravel-mssql-changes` will be documented in this file.
 
+## Unreleased Changes
+
+ - Sort ShowChanges output by Change Version (without grouping by Table)
+
 ## 1.0.0 - 2022-10-10
 
 Initial Release